refactor: 상태 유지 할 필요가 없는 클래스의 기능들 정적 메소드로 변경
