+++ src/main/java/baseball/Application.java
@@ -2,7 +2,7 @@ package baseball;
 
 public class Application {
     public static void main(String[] args) {
-        NumberBaseballGame numberBaseballGame = new NumberBaseballGame();
-        numberBaseballGame.start();
+        GameRunner gameRunner = new GameRunner();
+        gameRunner.start();
     }
 }
